% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_fct_utils.R
\name{check_all_modules_valid}
\alias{check_all_modules_valid}
\title{Check if all required modules are validated}
\usage{
check_all_modules_valid(session)
}
\arguments{
\item{session}{Shiny session object containing userData$reactiveValues}
}
\value{
logical indicating whether all required modules are validated
}
\description{
Checks whether all upstream modules have valid data for enabling measurement data entry
}
