% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_helpers.R
\name{copper_CREED_purpose_statement}
\alias{copper_CREED_purpose_statement}
\title{Load copper CREED purpose statement}
\usage{
copper_CREED_purpose_statement()
}
\value{
nested list of criteria and thresholds to meet them
}
\description{
Load a pre-written CREED purpose statement for a copper exposure assessment
}
