% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{create_metadata_tibble}
\alias{create_metadata_tibble}
\title{Create metadata tibble}
\usage{
create_metadata_tibble(metadata_list)
}
\arguments{
\item{metadata_list}{List containing metadata fields}
}
\value{
Tibble with Property and Value columns
}
\description{
Convert metadata list to tibble format suitable for Excel sheets
}
