% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dummy_data.R
\name{dummy_biota_tibble}
\alias{dummy_biota_tibble}
\title{Create Dummy Biota Tibble}
\usage{
dummy_biota_tibble()
}
\value{
A tibble with 1 row of dummy biota data
}
\description{
Creates a 1-row biota tibble with test data.
Inherits column structure from initialise_biota_tibble().
}
