% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dummy_data.R
\name{dummy_campaign_tibble}
\alias{dummy_campaign_tibble}
\title{Create Dummy Campaign Tibble}
\usage{
dummy_campaign_tibble()
}
\value{
A tibble with 1 row of dummy campaign data
}
\description{
Creates a 1-row campaign tibble with test data.
Inherits column structure from initialise_campaign_tibble().
}
