% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dummy_data.R
\name{dummy_references_tibble}
\alias{dummy_references_tibble}
\title{Create Dummy References Tibble}
\usage{
dummy_references_tibble()
}
\value{
A tibble with 1 row of dummy reference data
}
\description{
Creates a 1-row references tibble with test data.
Inherits column structure from initialise_references_tibble().
}
