% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dummy_data.R
\name{dummy_sites_tibble}
\alias{dummy_sites_tibble}
\title{Create Dummy Sites Tibble}
\usage{
dummy_sites_tibble()
}
\value{
A tibble with 2 rows of dummy site data
}
\description{
Creates a 2-row sites tibble with test data.
Inherits column structure from initialise_sites_tibble().
}
