% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{environ_compartment_subs_vocabulary}
\alias{environ_compartment_subs_vocabulary}
\title{Environmental Compartment Subs Controlled Vocabulary}
\usage{
environ_compartment_subs_vocabulary()
}
\value{
A character vector of environmental compartment subcategory options
}
\description{
Returns controlled vocabulary options for environmental compartment subcategories.
}
