% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{extract_campaign_name}
\alias{extract_campaign_name}
\title{Extract campaign name from campaign data for use in filenames}
\usage{
extract_campaign_name(campaign_data)
}
\arguments{
\item{campaign_data}{Data frame containing campaign information}
}
\value{
Character string of campaign name, or NULL if not found/invalid
}
\description{
Extract campaign name from campaign data for use in filenames
}
