% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{format_duration}
\alias{format_duration}
\title{Format a duration in seconds as a human-readable string}
\usage{
format_duration(seconds)
}
\arguments{
\item{seconds}{Numeric. Number of seconds to format.}
}
\value{
Character string.
}
\description{
Converts a numeric number of seconds into a readable string
such as "2 hours 13 minutes 51 seconds".
}
\examples{
format_duration(7451)   # "2 hours 4 minutes 11 seconds"
format_duration(30)     # "30 seconds"
}
