% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_all_schema_versions}
\alias{get_all_schema_versions}
\title{Get All Schema Versions}
\usage{
get_all_schema_versions()
}
\value{
A tibble with columns: schema_name, version
}
\description{
Returns a tibble with all schema names and their versions
}
