% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_biota_schema}
\alias{get_biota_schema}
\title{Get Biota Data Schema}
\usage{
get_biota_schema()
}
\value{
A list with version and schema tibble for biota data
}
\description{
Get Biota Data Schema
}
