% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_campaign_schema}
\alias{get_campaign_schema}
\title{Get Campaign Data Schema}
\usage{
get_campaign_schema()
}
\value{
A tibble with column specifications for campaign data
}
\description{
Get Campaign Data Schema
}
