% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_compartments_schema}
\alias{get_compartments_schema}
\title{Get Compartments Data Schema}
\usage{
get_compartments_schema()
}
\value{
A tibble with column specifications for compartments data
}
\description{
Get Compartments Data Schema
}
