% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_creed_scores_schema}
\alias{get_creed_scores_schema}
\title{Get CREED Scores Data Schema}
\usage{
get_creed_scores_schema()
}
\value{
A tibble with column specifications for CREED scores data
}
\description{
Get CREED Scores Data Schema
}
