% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{get_export_metadata}
\alias{get_export_metadata}
\title{Get export metadata}
\usage{
get_export_metadata(session = NULL)
}
\arguments{
\item{session}{Shiny session object. Required to access user data and client information.}
}
\value{
List containing export metadata fields (campaign_name, export_datetime, app_name, etc.)
}
\description{
Gather metadata about the current export session
}
