% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{get_git_commit}
\alias{get_git_commit}
\title{Get git commit hash}
\usage{
get_git_commit()
}
\value{
Character. Short git commit hash, or "Git hash not available" if retrieval fails
}
\description{
Retrieve the short git commit hash of the current repository state
}
