% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_methods_schema}
\alias{get_methods_schema}
\title{Get Methods Data Schema}
\usage{
get_methods_schema()
}
\value{
A tibble with column specifications for methods data
}
\description{
Get Methods Data Schema
}
