% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_parameters_schema}
\alias{get_parameters_schema}
\title{Get Parameters Data Schema}
\usage{
get_parameters_schema()
}
\value{
A tibble with column specifications for parameters data
}
\description{
Get Parameters Data Schema
}
