% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{get_reactiveValues_key}
\alias{get_reactiveValues_key}
\title{Get reactiveValues key for dataset type ----}
\usage{
get_reactiveValues_key(dataset_type)
}
\arguments{
\item{dataset_type}{Character string of dataset type}
}
\value{
Character string of reactiveValues key or NULL
}
\description{
Map dataset type to reactiveValues key
}
