% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_reference_schema}
\alias{get_reference_schema}
\title{Get Reference Data Schema}
\usage{
get_reference_schema()
}
\value{
A tibble with column specifications for reference data
}
\description{
Get Reference Data Schema
}
