% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_samples_schema}
\alias{get_samples_schema}
\title{Get Samples Data Schema}
\usage{
get_samples_schema()
}
\value{
A tibble with column specifications for samples data
}
\description{
Get Samples Data Schema
}
