% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data_safe.R
\name{get_session_data_safe}
\alias{get_session_data_safe}
\title{Safely retrieve data from session reactive values}
\usage{
get_session_data_safe(
  session,
  table_name,
  variable_name,
  fallback_value = "Unknown"
)
}
\arguments{
\item{session}{Shiny session object containing userData$reactiveValues}

\item{table_name}{Character string: name of the table in reactiveValues}

\item{variable_name}{Character string: name of the variable/column to retrieve}

\item{fallback_value}{Character string: value to return if conditions not met (default: "Unknown")}
}
\value{
Value from the specified table/variable if available, fallback_value otherwise
}
\description{
Safely retrieve data from session reactive values
}
