% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_sites_schema}
\alias{get_sites_schema}
\title{Get Sites Data Schema}
\usage{
get_sites_schema()
}
\value{
A list with version and schema tibble for sites data
}
\description{
Get Sites Data Schema
}
