% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utilities.R
\name{\%|truthy|\%}
\alias{\%|truthy|\%}
\title{isTruthy switch operator}
\usage{
first \%|truthy|\% second
}
\arguments{
\item{first}{The first argument, a value we want to use only if it isTruthy}

\item{second}{The second argument, a safe alternative if first isn't Truthy}
}
\value{
object types are unchanged
}
\description{
isTruthy switch operator
}
\details{
utilities


A simple switch operator based on purrr's %||% that returns the first
argument if it isTruthy (exists, not NULL/NA/"", etc.), and otherwise the
second argument.
}
