% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utilities.R
\name{hot_table_data_extract}
\alias{hot_table_data_extract}
\title{Extract Clean Data Frame from rhandsontable Input}
\usage{
hot_table_data_extract(hot_json)
}
\arguments{
\item{hot_json}{string of class "json" from rhandsontable input (typically from Shiny input)}
}
\value{
A cleaned data.frame
}
\description{
Converts rhandsontable input structure back to a clean R data frame,
handling common data type conversions and cleaning operations. Is this redundant 'cos of hot_to_r()?
}
