% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{import_module_table}
\alias{import_module_table}
\title{Import module dataset from CSV ----}
\usage{
import_module_table(csv_path, session)
}
\arguments{
\item{csv_path}{Path to CSV file}

\item{session}{Shiny session object}
}
\value{
List with success status and details
}
\description{
Import a single dataset and add to session reactiveValues
}
