% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{import_session_from_zip}
\alias{import_session_from_zip}
\title{Import data from exported ZIP file ----}
\usage{
import_session_from_zip(zip_path, session)
}
\arguments{
\item{zip_path}{Path to the ZIP file}

\item{session}{Shiny session object}
}
\value{
List with success status and any error messages
}
\description{
Import datasets from a ZIP file created by mod_export
}
