% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_fct_utils.R
\name{init_measurement_df}
\alias{init_measurement_df}
\title{Initialize empty measurement data frame}
\usage{
init_measurement_df()
}
\value{
tibble with zero rows but correct column structure for measurement data
}
\description{
Creates an empty tibble with the correct column structure for measurement data
}
