% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utilities.R
\name{isRelevant}
\alias{isRelevant}
\title{isRelevant}
\usage{
isRelevant(input)
}
\arguments{
\item{first}{The first argument, a value we want to use only if it isTruthy}

\item{second}{The second argument, a safe alternative if first isn't Truthy}
}
\value{
Boolean
}
\description{
Check if a categorical variable is "Not relevant" or "Not reported"
}
\details{
Check if a user has entered "Not relevant" or "Not reported" in selectInput
for the purpose of including conditional fields, validation, etc.
}
