% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_details.R
\name{mod_CREED_details_server}
\alias{mod_CREED_details_server}
\title{CREED_details Server Functions}
\usage{
mod_CREED_details_server(id)
}
\arguments{
\item{sessionData}{session$userData$reactiveValues object}
}
\value{
tibble of Dataset Details fields ("field") and pretty strings ("value")
}
\description{
summarise session user data into pretty strings for CREED "Dataset Details" reporting
}
