% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_details.R
\name{mod_CREED_details_ui}
\alias{mod_CREED_details_ui}
\title{CREED_details UI Function}
\usage{
mod_CREED_details_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
