% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_structured.R
\name{mod_structured_ui}
\alias{mod_structured_ui}
\alias{mod_structured_server}
\title{structured UI Function}
\usage{
mod_structured_ui(id)

mod_structured_server(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
