% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{object_to_text}
\alias{object_to_text}
\title{Convert object to human-readable text}
\usage{
object_to_text(obj, dataset_name = "unknown")
}
\arguments{
\item{obj}{The object to convert (character, list, ellmer_schema, or other object types)}

\item{dataset_name}{Character. Name of the dataset for header context}
}
\value{
Character vector suitable for use with writeLines()
}
\description{
Convert various R objects (character vectors, lists, S3/S4 objects) to
  readable text format suitable for writing to .txt files
}
