% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utilities.R
\name{printreactiveValues}
\alias{printreactiveValues}
\title{print content of reactiveValues object}
\usage{
printreactiveValues(data)
}
\arguments{
\item{data}{A reactiveValues object with named variables}
}
\value{
a string of variable names and values
}
\description{
Print a reactiveValues object, with each named variable and its value on a new
line
}
\examples{
reactiveValues(x = 1, y = 2) |> printreactiveValues()
}
