% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{protocol_options_vocabulary}
\alias{protocol_options_vocabulary}
\title{Protocol Options Data}
\usage{
protocol_options_vocabulary()
}
\value{
A data frame of protocol options
}
\description{
Returns protocol options data from parquet file.
}
