% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{read_metadata_txt}
\alias{read_metadata_txt}
\title{Read metadata from text file ----}
\usage{
read_metadata_txt(file_path)
}
\arguments{
\item{file_path}{Path to the metadata text file}
}
\value{
List with metadata or NULL if not found/readable
}
\description{
Read metadata from a human-readable text file
}
