% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_R6.R
\name{SamplesData_Table_R6_Generator}
\alias{SamplesData_Table_R6_Generator}
\title{Samples Data Table R6 Generator}
\description{
R6 class for managing sample metadata
}
\section{Super class}{
\code{\link[STOPeData:Data_Table_R6_Generator]{STOPeData::Data_Table_R6_Generator}} -> \code{SamplesData_Table_R6_Generator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SamplesData_Table_R6_Generator-new}{\code{SamplesData_Table_R6_Generator$new()}}
\item \href{#method-SamplesData_Table_R6_Generator-clone}{\code{SamplesData_Table_R6_Generator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="STOPeData" data-topic="Data_Table_R6_Generator" data-id="add_rows"><a href='../../STOPeData/html/Data_Table_R6_Generator.html#method-Data_Table_R6_Generator-add_rows'><code>STOPeData::Data_Table_R6_Generator$add_rows()</code></a></span></li>
<li><span class="pkg-link" data-pkg="STOPeData" data-topic="Data_Table_R6_Generator" data-id="validate_structure"><a href='../../STOPeData/html/Data_Table_R6_Generator.html#method-Data_Table_R6_Generator-validate_structure'><code>STOPeData::Data_Table_R6_Generator$validate_structure()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplesData_Table_R6_Generator-new"></a>}}
\if{latex}{\out{\hypertarget{method-SamplesData_Table_R6_Generator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamplesData_Table_R6_Generator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplesData_Table_R6_Generator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SamplesData_Table_R6_Generator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamplesData_Table_R6_Generator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
