% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{species_names_vocabulary}
\alias{species_names_vocabulary}
\title{Species Controlled Vocabulary}
\usage{
species_names_vocabulary()
}
\value{
A character vector of parameter type subcategory options
}
\description{
Returns species name options drom ECOTOX data
}
