% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOPeData-package.R
\docType{package}
\name{STOPeData-package}
\alias{STOPeData}
\alias{STOPeData-package}
\title{STOPeData: PKG_TITLE}
\description{
PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sawelch-NIVA/STOPeData}
  \item Report bugs at \url{https://github.com/sawelch-NIVA/STOPeData/issues}
}

}
\author{
\strong{Maintainer}: AUTHOR_FIRST AUTHOR_LAST \email{AUTHOR@MAIL.COM}

}
\keyword{internal}
