% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_protocols}
\alias{summarise_protocols}
\title{Summarise protocols data}
\usage{
summarise_protocols(methodsData, categories)
}
\arguments{
\item{methodsData}{The methods/protocols dataset. Use
\code{\link[eDataDRF:example_methods_tibble]{eDataDRF::example_methods_tibble()}} to generate an example input.}

\item{categories}{Character vector of protocol categories to include. Must be
values from \code{\link[eDataDRF:protocol_categories_vocabulary]{eDataDRF::protocol_categories_vocabulary()}} (e.g.,
\code{c("Sampling Protocol", "Analytical Protocol")}).}
}
\value{
Character string summarising protocols, or "Relevant data not found"
}
\description{
Creates a summary string of protocols by category
}
\examples{
summarise_protocols(
  eDataDRF::example_methods_tibble(),
  categories = c("Sampling Protocol", "Analytical Protocol")
)
}
\seealso{
\code{\link[eDataDRF:example_methods_tibble]{eDataDRF::example_methods_tibble()}},
\code{\link[eDataDRF:protocol_categories_vocabulary]{eDataDRF::protocol_categories_vocabulary()}},
\code{\link[eDataDRF:protocol_options_vocabulary]{eDataDRF::protocol_options_vocabulary()}}

Other summarise:
\code{\link[=calculate_coordinate_precision]{calculate_coordinate_precision()}},
\code{\link[=manual_completion_message]{manual_completion_message()}},
\code{\link[=summarise_CREED_details]{summarise_CREED_details()}},
\code{\link[=summarise_CREED_relevance]{summarise_CREED_relevance()}},
\code{\link[=summarise_CREED_reliability]{summarise_CREED_reliability()}},
\code{\link[=summarise_biota]{summarise_biota()}},
\code{\link[=summarise_compartments]{summarise_compartments()}},
\code{\link[=summarise_date_range]{summarise_date_range()}},
\code{\link[=summarise_lod_loq]{summarise_lod_loq()}},
\code{\link[=summarise_measured_units]{summarise_measured_units()}},
\code{\link[=summarise_multiple]{summarise_multiple()}},
\code{\link[=summarise_reference]{summarise_reference()}},
\code{\link[=summarise_sig_figs]{summarise_sig_figs()}},
\code{\link[=summarise_sites]{summarise_sites()}},
\code{\link[=summarise_uncertainty_comments]{summarise_uncertainty_comments()}}
}
\concept{summarise}
