% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_uncertainty_comments}
\alias{summarise_uncertainty_comments}
\title{Summarise uncertainty and measurement comments}
\usage{
summarise_uncertainty_comments(measurementsData)
}
\arguments{
\item{measurementsData}{The measurements dataset. Use
\code{\link[eDataDRF:example_measurements_tibble]{eDataDRF::example_measurements_tibble()}} to generate an example input.}
}
\value{
Character string summarising uncertainty info, or "Relevant data not found"
}
\description{
Creates a summary of uncertainty types and measurement comments
}
\examples{
summarise_uncertainty_comments(eDataDRF::example_measurements_tibble())
}
\seealso{
\code{\link[eDataDRF:example_measurements_tibble]{eDataDRF::example_measurements_tibble()}},
\code{\link[eDataDRF:uncertainty_types_vocabulary]{eDataDRF::uncertainty_types_vocabulary()}}

Other summarise:
\code{\link[=calculate_coordinate_precision]{calculate_coordinate_precision()}},
\code{\link[=manual_completion_message]{manual_completion_message()}},
\code{\link[=summarise_CREED_details]{summarise_CREED_details()}},
\code{\link[=summarise_CREED_relevance]{summarise_CREED_relevance()}},
\code{\link[=summarise_CREED_reliability]{summarise_CREED_reliability()}},
\code{\link[=summarise_biota]{summarise_biota()}},
\code{\link[=summarise_compartments]{summarise_compartments()}},
\code{\link[=summarise_date_range]{summarise_date_range()}},
\code{\link[=summarise_lod_loq]{summarise_lod_loq()}},
\code{\link[=summarise_measured_units]{summarise_measured_units()}},
\code{\link[=summarise_multiple]{summarise_multiple()}},
\code{\link[=summarise_protocols]{summarise_protocols()}},
\code{\link[=summarise_reference]{summarise_reference()}},
\code{\link[=summarise_sig_figs]{summarise_sig_figs()}},
\code{\link[=summarise_sites]{summarise_sites()}}
}
\concept{summarise}
