% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{write_metadata_yaml}
\alias{write_metadata_yaml}
\title{Write metadata as a YAML file}
\usage{
write_metadata_yaml(metadata_list, file_path)
}
\arguments{
\item{metadata_list}{List. The metadata to write.}

\item{file_path}{Character. Path where to write the YAML file.}
}
\value{
NULL (invisibly). File is written to disk as a side effect.
}
\description{
Serialises a metadata list to a YAML file using the yaml package.
}
\examples{
\dontrun{
  write_metadata_yaml(list(session = list(user = "Ole")), tempfile(fileext = ".yaml"))
}
}
