% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{calculate_coordinate_precision}
\alias{calculate_coordinate_precision}
\title{Calculate coordinate precision}
\usage{
calculate_coordinate_precision(latitude, longitude)
}
\arguments{
\item{latitude}{Numeric vector of latitudes}

\item{longitude}{Numeric vector of longitudes}
}
\value{
Integer representing minimum decimal places, or "N/A" if no valid coords
}
\description{
Determines the minimum decimal places across latitude and longitude
}
