% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_helpers.R
\name{collect_CREED_data}
\alias{collect_CREED_data}
\title{Collect CREED Scores from Input}
\usage{
collect_CREED_data(criteria_config, input)
}
\arguments{
\item{criteria_config}{the CREED criteria for either reliability or relevance}

\item{input}{the shiny input object for the module}
}
\description{
Collects CREED criterion scores, relevant data, and limitations
from Shiny inputs into a standardised tibble. Works for both reliability and
relevance criteria.
}
