% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compartments_fct_.R
\name{create_compartment_combination}
\alias{create_compartment_combination}
\title{Create new compartment combination row}
\usage{
create_compartment_combination(compartment, sub_compartment, category)
}
\arguments{
\item{compartment}{Character string specifying the environmental compartment (\code{\link[eDataDRF:environ_compartments_vocabulary]{eDataDRF::environ_compartments_vocabulary()}})}

\item{sub_compartment}{Character string specifying the sub-compartment (\code{\link[eDataDRF:environ_compartments_sub_vocabulary]{eDataDRF::environ_compartments_sub_vocabulary()}})}

\item{category}{Character string specifying the measured category (\code{\link[eDataDRF:measured_category_vocabulary]{eDataDRF::measured_category_vocabulary()}})}
}
\value{
tibble with one row containing the specified compartment information
}
\description{
Creates a single-row tibble with specified compartment information,
using the standardised structure from \code{\link[eDataDRF:initialise_campaign_tibble]{eDataDRF::initialise_campaign_tibble()}}.
Function will return an error if compartment, sub_compartment, or category are
not of the correct variable type, but it doesn't check for vocabulary
}
\examples{
create_compartment_combination("Aquatic", "Surface water", "External")
create_compartment_combination("Biota", "Biota, Aquatic", "Internal")
}
\seealso{
\code{\link[eDataDRF:initialise_campaign_tibble]{eDataDRF::initialise_campaign_tibble()}}

Other create: 
\code{\link{create_dummy_session_data}()},
\code{\link{create_existing_parameter}()},
\code{\link{create_new_parameter}()},
\code{\link{create_new_site}()}
}
\concept{create}
