% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compartments_fct_.R
\name{create_compartment_combination}
\alias{create_compartment_combination}
\title{Create new compartment combination row}
\usage{
create_compartment_combination(compartment, sub_compartment, category)
}
\arguments{
\item{compartment}{Character string specifying the environmental compartment}

\item{sub_compartment}{Character string specifying the sub-compartment}

\item{category}{Character string specifying the measured category}
}
\value{
tibble with one row containing the specified compartment information
}
\description{
Creates a single-row tibble with specified compartment information,
using the standardised structure from initialise_compartments_tibble().
}
\seealso{
\code{\link{initialise_compartments_tibble}}
}
