% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameters_fct.R
\name{create_new_parameter}
\alias{create_new_parameter}
\title{Create new blank parameter row}
\usage{
create_new_parameter(param_type, entered_by)
}
\arguments{
\item{param_type}{Character string specifying the parameter type}

\item{entered_by}{Character string specifying who entered the parameter}
}
\value{
tibble with blank parameter template
}
\description{
Creates a single-row template tibble with default values for a new parameter.
}
\examples{
create_new_parameter("Chemical", "Jane Smith")
}
\seealso{
\code{\link{initialise_parameters_tibble}}

Other create: 
\code{\link{create_compartment_combination}()},
\code{\link{create_dummy_session_data}()},
\code{\link{create_existing_parameter}()},
\code{\link{create_new_site}()}
}
\concept{create}
