% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameters_fct_.R
\name{create_new_parameter}
\alias{create_new_parameter}
\title{Create new blank parameter row}
\usage{
create_new_parameter(param_type)
}
\arguments{
\item{param_type}{Character string specifying the parameter type}
}
\value{
Data.frame with blank parameter template
}
\description{
Create new blank parameter row
}
