% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sites_fct.R
\name{create_new_site}
\alias{create_new_site}
\title{Create new site row with defaults}
\usage{
create_new_site(site_number = 1, base_code = "", username)
}
\arguments{
\item{site_number}{Integer. Site number for generating the site code. Default is 1.}

\item{base_code}{Character. Base code prefix for the site code. If empty or NULL,
defaults to "SITE_". Default is "".}

\item{username}{Character. Username of the person entering the site data.}
}
\value{
A tibble with one row containing the new site with default values.
}
\description{
Create new site row with defaults
}
\examples{
\dontrun{
create_new_site(site_number = 1, base_code = "LAKE_", username = "jdoe")
create_new_site(site_number = 42, username = "jsmith")
}

}
