% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_samples_fct.R
\name{generate_sample_id_with_components}
\alias{generate_sample_id_with_components}
\title{Generate Sample ID with Components ----}
\usage{
generate_sample_id_with_components(
  site_code,
  parameter_name,
  environ_compartment,
  environ_compartment_sub,
  date,
  subsample = 1
)
}
\arguments{
\item{site_code}{Site code (vectorized)}

\item{parameter_name}{Parameter name (vectorized)}

\item{environ_compartment}{Environmental compartment (vectorized)}

\item{environ_compartment_sub}{Environmental sub-compartment (vectorized)}

\item{date}{Sampling date (vectorized)}

\item{subsample}{subsample}
}
\description{
Generate Sample ID with Components ----
}
