% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{get_measurements_schema}
\alias{get_measurements_schema}
\title{Get Measurements Data Schema}
\usage{
get_measurements_schema()
}
\value{
A list with version and schema tibble for measurements data
}
\description{
Get Measurements Data Schema
}
