% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameters_fct.R
\name{get_parameters_filtered}
\alias{get_parameters_filtered}
\title{Get parameter names filtered by type and optionally subtype}
\usage{
get_parameters_filtered(
  param_type,
  param_subtype = "Show all",
  dummy_parameters,
  session_parameters = NULL
)
}
\arguments{
\item{param_type}{Character string specifying the parameter type}

\item{param_subtype}{Character string specifying the parameter subtype (optional)}

\item{dummy_parameters}{Dataframe containing base parameters}

\item{session_parameters}{Optional list containing session-specific parameters}
}
\value{
Character vector of parameter names
}
\description{
Get parameter names filtered by type and optionally subtype
}
