% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameters_fct.R
\name{get_parameters_of_types}
\alias{get_parameters_of_types}
\title{Get parameter names for a specific parameter type}
\usage{
get_parameters_of_types(
  param_type,
  dummy_parameters,
  session_parameters = NULL
)
}
\arguments{
\item{param_type}{Character string specifying the parameter type}

\item{dummy_parameters}{Dataframe containing base parameters}

\item{session_parameters}{Optional list containing session-specific parameters}
}
\value{
Character vector of parameter names
}
\description{
Get parameter names for a specific parameter type
}
