% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_utilities.R
\name{info_accordion}
\alias{info_accordion}
\title{Create a collapsible single-panel accordion containing a markdown file}
\usage{
info_accordion(title = "Instructions", content_file)
}
\arguments{
\item{title}{the desired title of the accordion panel}

\item{content_file}{the path to a markdown file}
}
\value{
a bslib::accordion html element
}
\description{
Create a collapsible single-panel accordion containing a markdown file
}
