% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{manual_completion_message}
\alias{manual_completion_message}
\title{Return manual completion message}
\usage{
manual_completion_message()
}
\value{
Character string indicating manual completion required
}
\description{
Returns a standard message for fields not collected by the app
}
\examples{
manual_completion_message()
}
\seealso{
Other summarise: 
\code{\link{calculate_coordinate_precision}()},
\code{\link{summarise_CREED_details}()},
\code{\link{summarise_CREED_relevance}()},
\code{\link{summarise_CREED_reliability}()},
\code{\link{summarise_biota}()},
\code{\link{summarise_compartments}()},
\code{\link{summarise_date_range}()},
\code{\link{summarise_lod_loq}()},
\code{\link{summarise_measured_units}()},
\code{\link{summarise_multiple}()},
\code{\link{summarise_protocols}()},
\code{\link{summarise_reference}()},
\code{\link{summarise_sig_figs}()},
\code{\link{summarise_sites}()},
\code{\link{summarise_uncertainty_comments}()}
}
\concept{summarise}
