% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{manual_completion_message}
\alias{manual_completion_message}
\title{Return manual completion message}
\usage{
manual_completion_message()
}
\value{
Character string indicating manual completion required
}
\description{
Returns a standard message for fields not collected by the app
}
