% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{summarise_biota}
\alias{summarise_biota}
\title{Summarise biota data}
\usage{
summarise_biota(
  biotaData,
  SPECIES_GROUP = FALSE,
  SAMPLE_SPECIES = FALSE,
  SAMPLE_TISSUE = FALSE,
  SAMPLE_SPECIES_LIFESTAGE = FALSE,
  SAMPLE_SPECIES_GENDER = FALSE
)
}
\arguments{
\item{biotaData}{The biota dataset}

\item{SPECIES_GROUP}{Logical. Include species group summary?}

\item{SAMPLE_SPECIES}{Logical. Include sample species summary?}

\item{SAMPLE_TISSUE}{Logical. Include tissue type summary?}

\item{SAMPLE_SPECIES_LIFESTAGE}{Logical. Include life stage summary?}

\item{SAMPLE_SPECIES_GENDER}{Logical. Include gender summary?}
}
\value{
Character string summarising biota, or "Relevant data not found"
}
\description{
Creates a summary string of biota including species, tissues, and life stages
}
