% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_compartments}
\alias{summarise_compartments}
\title{Summarise compartments data}
\usage{
summarise_compartments(compartmentsData)
}
\arguments{
\item{compartmentsData}{The compartments dataset}
}
\value{
Character string summarising compartments, or "Relevant data not found"
}
\description{
Creates a summary string of compartments from session data
}
\examples{
comps <- data.frame(
  ENVIRON_COMPARTMENT = c("Water", "Sediment"),
  ENVIRON_COMPARTMENT_SUB = c("Surface water", NA)
)
summarise_compartments(comps)
}
\seealso{
Other summarise: 
\code{\link{calculate_coordinate_precision}()},
\code{\link{manual_completion_message}()},
\code{\link{summarise_CREED_details}()},
\code{\link{summarise_CREED_relevance}()},
\code{\link{summarise_CREED_reliability}()},
\code{\link{summarise_biota}()},
\code{\link{summarise_date_range}()},
\code{\link{summarise_lod_loq}()},
\code{\link{summarise_measured_units}()},
\code{\link{summarise_multiple}()},
\code{\link{summarise_protocols}()},
\code{\link{summarise_reference}()},
\code{\link{summarise_sig_figs}()},
\code{\link{summarise_sites}()},
\code{\link{summarise_uncertainty_comments}()}
}
\concept{summarise}
