% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_compartments}
\alias{summarise_compartments}
\title{Summarise compartments data}
\usage{
summarise_compartments(compartmentsData)
}
\arguments{
\item{compartmentsData}{The compartments dataset}
}
\value{
Character string summarising compartments, or "Relevant data not found"
}
\description{
Creates a summary string of compartments from session data
}
