% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_helpers.R
\name{summarise_CREED_details}
\alias{summarise_CREED_details}
\title{summarise user-entered data for the CREED dataset details reporting}
\usage{
summarise_CREED_details(sessionData)
}
\arguments{
\item{sessionData}{session$userData$reactiveValues object}
}
\value{
tibble of Dataset Details fields ("field") and pretty strings ("value")
}
\description{
summarise session user data into pretty strings for CREED "Dataset Details" reporting
}
