% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_helpers.R
\name{summarise_CREED_relevance}
\alias{summarise_CREED_relevance}
\title{summarise user-entered data for the CREED relevance criteria reporting}
\usage{
summarise_CREED_relevance(sessionData)
}
\arguments{
\item{sessionData}{session$userData$reactiveValues object}
}
\value{
tibble of Relevance criteria fields ("field") and pretty strings ("value")
}
\description{
summarise session user data into pretty strings for CREED relevance criteria
}
\examples{
\dontrun{
  # sessionData is the reactiveValues object from the Shiny app
  relevance <- summarise_CREED_relevance(session$userData)
  relevance
}
}
\seealso{
Other summarise: 
\code{\link{calculate_coordinate_precision}()},
\code{\link{manual_completion_message}()},
\code{\link{summarise_CREED_details}()},
\code{\link{summarise_CREED_reliability}()},
\code{\link{summarise_biota}()},
\code{\link{summarise_compartments}()},
\code{\link{summarise_date_range}()},
\code{\link{summarise_lod_loq}()},
\code{\link{summarise_measured_units}()},
\code{\link{summarise_multiple}()},
\code{\link{summarise_protocols}()},
\code{\link{summarise_reference}()},
\code{\link{summarise_sig_figs}()},
\code{\link{summarise_sites}()},
\code{\link{summarise_uncertainty_comments}()}
}
\concept{summarise}
