% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_helpers.R
\name{summarise_CREED_relevance}
\alias{summarise_CREED_relevance}
\title{summarise user-entered data for the CREED relevance criteria reporting}
\usage{
summarise_CREED_relevance(sessionData)
}
\arguments{
\item{sessionData}{session$userData$reactiveValues object}
}
\value{
tibble of Relevance criteria fields ("field") and pretty strings ("value")
}
\description{
summarise session user data into pretty strings for CREED relevance criteria
}
