% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_date_range}
\alias{summarise_date_range}
\title{Calculate Date Range}
\usage{
summarise_date_range(dates)
}
\arguments{
\item{dates}{Vector of dates}
}
\value{
Character string with date range
}
\description{
Creates a date range string from vector of dates
}
