% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_lod_loq}
\alias{summarise_lod_loq}
\title{Summarise LOD/LOQ data}
\usage{
summarise_lod_loq(measurementsData)
}
\arguments{
\item{measurementsData}{The measurements dataset}
}
\value{
Character string summarising LOD/LOQ, or "Relevant data not found"
}
\description{
Creates a summary string of limit of detection and quantification values
}
