% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_measured_units}
\alias{summarise_measured_units}
\title{Generate Units Summary by Parameter}
\usage{
summarise_measured_units(measurement_data, parameters_data)
}
\arguments{
\item{measurement_data}{Measurement data frame}

\item{parameters_data}{Parameters data frame}
}
\value{
Character string with units grouped by parameter
}
\description{
Creates a summary of measurement units grouped by parameter name
}
