% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_protocols}
\alias{summarise_protocols}
\title{Summarise protocols data}
\usage{
summarise_protocols(methodsData, categories)
}
\arguments{
\item{methodsData}{The methods/protocols dataset}

\item{categories}{Character vector of protocol categories to include (e.g., c("Sampling Protocol", "Analytical Protocol"))}
}
\value{
Character string summarising protocols, or "Relevant data not found"
}
\description{
Creates a summary string of protocols by category
}
