% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_reference}
\alias{summarise_reference}
\title{Summarise mod_reference data into a single string}
\usage{
summarise_reference(ref_data)
}
\arguments{
\item{ref_data}{Reference data frame}
}
\value{
Character string with formatted reference
}
\description{
Creates a formatted bibliographic reference from reference data
}
\examples{
ref <- data.frame(
  AUTHOR = "Smith J.",
  YEAR = 2022,
  TITLE = "Marine pollution monitoring",
  PERIODICAL_JOURNAL = "Marine Pollution Bulletin",
  PUBLISHER = NA,
  DOI = "10.1000/xyz123"
)
summarise_reference(ref)
}
\seealso{
Other summarise: 
\code{\link{calculate_coordinate_precision}()},
\code{\link{manual_completion_message}()},
\code{\link{summarise_CREED_details}()},
\code{\link{summarise_CREED_relevance}()},
\code{\link{summarise_CREED_reliability}()},
\code{\link{summarise_biota}()},
\code{\link{summarise_compartments}()},
\code{\link{summarise_date_range}()},
\code{\link{summarise_lod_loq}()},
\code{\link{summarise_measured_units}()},
\code{\link{summarise_multiple}()},
\code{\link{summarise_protocols}()},
\code{\link{summarise_sig_figs}()},
\code{\link{summarise_sites}()},
\code{\link{summarise_uncertainty_comments}()}
}
\concept{summarise}
