% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_reference}
\alias{summarise_reference}
\title{Summarise mod_reference data into a single string}
\usage{
summarise_reference(ref_data)
}
\arguments{
\item{ref_data}{Reference data frame}
}
\value{
Character string with formatted reference
}
\description{
Creates a formatted bibliographic reference from reference data
}
