% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_sig_figs}
\alias{summarise_sig_figs}
\title{Summarise significant figures in numeric column}
\usage{
summarise_sig_figs(values)
}
\arguments{
\item{values}{Numeric vector to analyze}
}
\value{
Character string describing significant figures, or "Relevant data not found"
}
\description{
Calculates and summarises the significant figures in a numeric vector
}
