% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_helpers.R
\name{summarise_sites}
\alias{summarise_sites}
\title{Summarize sites data}
\usage{
summarise_sites(sitesData)
}
\arguments{
\item{sitesData}{The sites dataset}
}
\value{
Character string summarizing sites, or "Relevant data not found"
}
\description{
Creates a summary string of sites including countries and areas
}
