% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_sites}
\alias{summarise_sites}
\title{Summarise sites data}
\usage{
summarise_sites(
  sitesData,
  COUNTRY_ISO = FALSE,
  OCEAN_IHO = FALSE,
  SITE_GEOGRAPHIC_FEATURE = FALSE,
  SITE_GEOGRAPHIC_FEATURE_SUB = FALSE,
  PRECISION = FALSE
)
}
\arguments{
\item{sitesData}{The sites dataset}

\item{COUNTRY_ISO}{Logical. Include country summary?}

\item{OCEAN_IHO}{Logical. Include area summary?}

\item{SITE_GEOGRAPHIC_FEATURE}{Logical. Include geographic feature summary?}

\item{SITE_GEOGRAPHIC_FEATURE_SUB}{Logical. Include geographic feature sub summary?}

\item{PRECISION}{Logical. Include coordinate precision?}
}
\value{
Character string summarising sites, or "Relevant data not found"
}
\description{
Creates a summary string of sites including countries and areas
}
