% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CREED_fct_summarise.R
\name{summarise_uncertainty_comments}
\alias{summarise_uncertainty_comments}
\title{Summarise uncertainty and measurement comments}
\usage{
summarise_uncertainty_comments(measurementsData)
}
\arguments{
\item{measurementsData}{The measurements dataset}
}
\value{
Character string summarising uncertainty info, or "Relevant data not found"
}
\description{
Creates a summary of uncertainty types and measurement comments
}
