% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{validate_dataset_structure}
\alias{validate_dataset_structure}
\title{Validate dataset structure ----}
\usage{
validate_dataset_structure(data, dataset_type)
}
\arguments{
\item{data}{Tibble/data.frame to validate}

\item{dataset_type}{Character string of dataset type}
}
\value{
List with valid (logical) and message (character)
}
\description{
Basic validation of dataset structure
}
