% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download.R
\name{write_metadata_txt}
\alias{write_metadata_txt}
\title{Create readable metadata text file ----}
\usage{
write_metadata_txt(metadata_list, file_path)
}
\arguments{
\item{metadata_list}{List containing metadata}

\item{file_path}{Path where to write the metadata file}
}
\description{
Create a human-readable text file with export metadata
}
