% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_zenodo.R
\docType{data}
\name{zenodo_licenses}
\alias{zenodo_licenses}
\title{Zenodo license options tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 444 rows and 10 columns.
}
\usage{
zenodo_licenses
}
\description{
Returns a tibble of Zenodo license options as of March 2026.
Includes \code{id}, \code{title}, \code{description}, \code{url}, \code{schema}, \code{osi_approved}, \code{revision_id},
\code{created}, \code{updated}, and \code{popular}, a custom boolean highlighting the most commonly
used licenses.
}
\examples{
# Show only popular licenses
zenodo_licenses[zenodo_licenses$popular, c("id", "title")]
}
\seealso{
Other zenodo: 
\code{\link{generate_zenodo_readme}()},
\code{\link{zenodo_resource_types}}
}
\concept{zenodo}
\keyword{datasets}
