% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_zenodo.R
\docType{data}
\name{zenodo_licenses}
\alias{zenodo_licenses}
\title{Zenodo license options tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 444 rows and 10 columns.
}
\usage{
zenodo_licenses
}
\description{
Returns a tibble of Zenodo license options as of March 2026.
Includes `id`, `title`, `description`, `url`, `schema`, `osi_approved`, `revision_id`,
`created`, `updated`, and `popular`, a custom boolean highlighting the most commonly
used licenses.
}
\keyword{datasets}
