% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_zenodo.R
\docType{data}
\name{zenodo_resource_types}
\alias{zenodo_resource_types}
\title{Zenodo resource type options vector}
\format{
An object of class \code{character} of length 5.
}
\usage{
zenodo_resource_types
}
\description{
Named character vector of Zenodo resource types.
  Names are display labels; values are Zenodo resource type identifiers.
  See \url{https://help.zenodo.org/docs/deposit/describe-records/resource-type/}
  for more details. Filtered to values likely to be used in this application.
}
\keyword{datasets}
